requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
yfinance==0.2.18
fastapi==0.103.1
uvicorn==0.23.2
pydantic==1.10.13
dvc==3.15.2
mlflow==2.6.0
python-dotenv==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
jupyter==1.0.0
requests==2.31.0
python-multipart==0.0.6
xgboost>=1.7.0